nixos/acme: Add emailFromEnvironment option
Allows the loading of the email address to use from environment
variables. This was recently allowed in lego itself, but there were
assertions and technical limitations preventing this in the module.

Note that the accountHash still utilises data.email explicitly. This is
a conscious choice. The account key will be uniquely named
within the directory, and there is an explicit check for its existence
in the renewal bash script. Users are most likely always using the same
email across multiple certs anyway.

Closes NixOS/nixpkgs#370825